+++ pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>tz.go.moh.him</groupId>
-    <artifactId>hfr-mediator</artifactId>
+    <artifactId>elmis-mediator-e9</artifactId>
     <version>0.1.0</version>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -41,7 +41,7 @@
                             </descriptorRefs>
                             <archive>
                                 <manifest>
-                                    <mainClass>tz.go.moh.him.hfr.mediator.MediatorMain</mainClass>
+                                    <mainClass>tz.go.moh.him.elmis.mediator.e9.MediatorMain</mainClass>
                                 </manifest>
                             </archive>
                         </configuration>
