+++ src/main/java/tz/go/moh/him/elmis/mediator/e9/MediatorMain.java
@@ -1,4 +1,4 @@
-package tz.go.moh.him.hfr.mediator;
+package tz.go.moh.him.elmis.mediator.e9;
 
 import akka.actor.ActorSystem;
 import akka.event.Logging;
